requirements.txt
django
djangorestframework
pytest-django
django-filter